requirements.txt
fastapi
uvicorn
confluent_kafka
fastapi_socketio
python-socketio
asyncio
